Separate Queries, QueryLoader, and types into separate modules.
aiosql.py was getting a bit heavy.
