+++ aiosql/__init__.py
@@ -1,3 +1,4 @@
-from .aiosql import from_path, from_str, SQLOperationType
+from .aiosql import from_path, from_str
+from .exceptions import SQLParseException, SQLLoadException
 
-__all__ = ["from_path", "from_str", "SQLOperationType"]
+__all__ = ["from_path", "from_str", "SQLParseException", "SQLLoadException"]
